requirements.txt
# Core dependencies
google-generativeai==0.8.3
google-ai-generativelanguage==0.6.10

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Data processing
numpy==2.1.2
pandas==2.2.3
python-dateutil==2.9.0

# News sources
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
aiohttp==3.10.10

# Vector operations and similarity
scikit-learn==1.5.2
chromadb==0.5.11

# Utilities
python-dotenv==1.0.1
tqdm==4.66.5
cachetools==5.5.0

# Frontend serving
jinja2==3.1.4

# Deep research and search
gpt-researcher==0.9.8
tavily-python==0.5.0

